+++ test/endpoint.js
@@ -30,6 +30,7 @@ describe('metrics', function() {
 describe('phonetics', function() {
   require('./phonetics/caverphone.js');
   require('./phonetics/cologne.js');
+  require('./phonetics/daitch-mokotoff.js');
   require('./phonetics/doubleMetaphone.js');
   require('./phonetics/metaphone.js');
   require('./phonetics/mra.js');